+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Allow VerifyUserEmailView get_redirect_url function to accept an extra string.
 
+* Allow user login after email verification providing a setting is true in an apps settings file.
+  This works in Django 1.10.
+
 ## 16.0.1
 
 * Fix email verification when `LOGIN_URL` is a url name.